+++ js/data.js
@@ -1,9 +1,10 @@
+import { getRandomInteger, createRandomIdGenerator } from './util.js';
 
 const URL_MAX = 25;
 const AVATAR_MAX_COUNT = 6;
 const LIKES_MIN_COUNT = 15;
 const LIKES_MAX_COUNT = 200;
-const FOTOS_COUNT = 25;
+const PHOTOS_COUNT = 25;
 const NAMES = [
   'Елена',
   'Кирилл',
@@ -42,39 +43,11 @@ const DESCRIPTION_LINES = [
   'Наряжаем елку'
 ]
 
-const getRandomInteger = (min, max) => {
-  const lower = Math.ceil(Math.min(min, max));
-  const upper = Math.floor(Math.max(min, max));
-  const result = Math.random() * (upper - lower + 1) + lower;
-  return Math.floor(result);
-};
-
-const createRandomIdGenerator = (min, max) => {
-  const previousValues = [];
-  return function () {
-    let currentValue = getRandomInteger(min, max);
-    if (previousValues.length >= (max - min + 1)) {
-
-      return null;
-    }
-    while (previousValues.includes(currentValue)) {
-      currentValue = getRandomInteger(min, max);
-    }
-    previousValues.push(currentValue);
-    return currentValue;
-  };
-};
-
 const generateId = createRandomIdGenerator(1, 25);
 const generateUrl = createRandomIdGenerator(1, URL_MAX);
 const generateIdComment = createRandomIdGenerator(1, 500);
-
-
-// console.log(getRandomInteger(1,25));
-
-
 const getRandomElementArray = (elements) => elements[getRandomInteger(0, elements.length - 1)];
-// console.log(getRandomElementArray(COMMENT_LINES))
+
 
 const createPostCard = () => {
   return {
@@ -90,6 +63,6 @@ const createPostCard = () => {
     }
   };
 };
-const PostCard = Array.from({length: FOTOS_COUNT}, createPostCard);
+const PostCard = Array.from({length: PHOTOS_COUNT}, createPostCard);
 
 export { PostCard };
